added factory provider
