--- spec/id_factory.go
@@ -1,15 +1,0 @@
-package spec
-
-// IDType represents some kind of configuration for ID creation.
-type IDType int
-
-// IDFactory creates pseudo random hash generation used for ID assignment.
-type IDFactory interface {
-	// New tries to create a new object ID using the configured ID type. The
-	// returned error might be caused by timeouts reached during the ID creation.
-	New() (ObjectID, error)
-
-	// WithType tries to create a new object ID using the given ID type. The
-	// returned error might be caused by timeouts reached during the ID creation.
-	WithType(idType IDType) (ObjectID, error)
-}
